+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Your contribution here.
 * [#82](https://github.com/ruby-grape/grape-swagger-rails/pull/82): Fixed api_key_default_value - [@konto-andrzeja](https://github.com/konto-andrzeja).
+* [#84](https://github.com/ruby-grape/grape-swagger-rails/pull/84): Added the token api_auth type - [@Jack12816](https://github.com/Jack12816).
 
 ### 0.3.0 (September 22, 2016)